pyproject.toml
[tool.poetry]
name = "homework-web-11"
version = "0.1.0"
description = ""
authors = ["ivan-markovskyi <ivan.markovskyi.in@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
sqlalchemy = "^2.0.25"
email-validator = "^2.1.0.post1"
pydantic-extra-types = "^2.4.1"
phonenumbers = "^8.13.28"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
